requirements.txt
pandas>=2.2
pyarrow>=15
openpyxl>=3.1
xlsxwriter>=3.2
streamlit>=1.33
PyYAML>=6.0
pytest>=8.2 